:recycle: Refactor tasks to use ListTasksComponent
